New view. Updated api imports
